refactor DaySlot to DayColumn (cosmetic)
